#33 wip move layout to shared components
